# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GuimaAPI",
        "description": "API para gerenciamento de alunos, treino, dieta, aulas e avaliações físicas.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/alunos": {
            "get": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Listar todos os alunos",
                "operationId": "9bf7261a9f1659e8fcc37d819026ea4d",
                "responses": {
                    "200": {
                        "description": "Retorna a lista de alunos"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/v1/alunos/{id}": {
            "get": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Mostrar um aluno específico por id",
                "operationId": "a5827235a86cf7f6101a78ea358640f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buscar por id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o aluno específico por id"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Deletar um aluno específico por id",
                "description": "Deletar um aluno específico por id",
                "operationId": "8e96d86e16ac967d998a8b307ad3e457",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do aluno",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Aluno deletado com sucesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Atualizar um aluno específico por id",
                "operationId": "8f42457b23308896fc76829f32242d8a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Atualizar por id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-12-31"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Rua das Flores, 123"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joao@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "plano": {
                                        "type": "enum",
                                        "example": "mensal"
                                    },
                                    "vencimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-31"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o aluno atualizado por id"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}